Impala Image Flipped
Impala ("impala4.JPG") image was flipped, and CSS flipping code to flip the unflipped image was removed, which was necessary before the actual image was flipped. This makes it less likely for client-side bugs to occur. After the flipping CSS code in "index.html" was updated as mentioned above, the old "index.html.zip" ZIP archive file was removed and replaced with a new one, containing the updated "index.html" file.